+++ examples/basic/src/model-function/embed-example.ts
@@ -5,7 +5,7 @@ dotenv.config();
 
 async function main() {
   const embedding = await embed({
-    model: openai.TextEmbedder({ model: "text-embedding-ada-002" }),
+    model: openai.TextEmbedder({ model: "text-embedding-3-small" }),
     value: "At first, Nox didn't know what to do with the pup.",
   });