+++ examples/basic/src/model-provider/huggingface/huggingface-embed-many-example.ts
@@ -7,7 +7,7 @@ async function main() {
   const embeddings = await embedMany({
     model: huggingface.TextEmbedder({
       model: "intfloat/e5-base-v2",
-      embeddingDimensions: 768,
+      dimensions: 768,
     }),
     values: [
       "At first, Nox didn't know what to do with the pup.",