+++ examples/basic/src/model-provider/openai/openai-embed-many-custom-dimensions-example.ts
@@ -1,0 +1,22 @@
+import dotenv from "dotenv";
+import { embedMany, openai } from "modelfusion";
+
+dotenv.config();
+
+async function main() {
+  const embeddings = await embedMany({
+    model: openai.TextEmbedder({
+      model: "text-embedding-3-large",
+      dimensions: 256,
+    }),
+    values: [
+      "At first, Nox didn't know what to do with the pup.",
+      "He keenly observed and absorbed everything around him, " +
+        "from the birds in the sky to the trees in the forest.",
+    ],
+  });
+
+  console.log(embeddings);
+}
+
+main().catch(console.error);